# gestion-soutenance-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Soutenances",
        "description": "Documentation OpenAPI pour le backend Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/admin/etudiants": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un étudiant",
                "operationId": "37fc052773dec1eca97dbc3dcd5f5632",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "mot_de_passe",
                                    "filiere",
                                    "type_stage"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Étudiant créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/etudiants/{etudiant}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Mettre à jour un étudiant",
                "operationId": "4d151a44819abf3afb1a1974bad4609e",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un étudiant",
                "operationId": "1d5f4c3f66e30726d68fe646c958369b",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/etudiants/{etudiant}/affectations": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Affecter encadrant/rapporteur",
                "operationId": "f1e24967c13080d0651b0333d5b99026",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "encadrant_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "rapporteur_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/professeurs": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un professeur",
                "operationId": "cf1e2c50cb63465e820288676b2ddcf8",
                "responses": {
                    "201": {
                        "description": "Professeur créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/professeurs/{professeur}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Mettre à jour un professeur",
                "operationId": "1634138ac10fda83bf891647626ec898",
                "parameters": [
                    {
                        "name": "professeur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un professeur",
                "operationId": "ec02677e5ec717ba072a6bce027fa0e9",
                "parameters": [
                    {
                        "name": "professeur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/etudiant/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion étudiant",
                "operationId": "a1bfb91c38b30135d1cf57cd68589f48",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token et profil"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/auth/professeur/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion professeur",
                "operationId": "5c060268c7ada069362ce915778226f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token et profil"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/auth/admin/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion administrateur",
                "operationId": "574417459337e78a0e6bec1a1dfe23a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token et profil"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Saisir uniquement le jeton d'accès",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}